requirements.txt
faker
pandas
pymongo
python-dotenv
schedule
streamlit
